+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ## 2.1 branch
 ### 2.1.8
 * support for some BBCode;
-* with the "<!-- read-more -->" tag, you can indicate manually where to truncate a text.
+* with the "<!-- read-more -->" tag, you can indicate manually where to truncate a text;
+* added Facebook's tag.
 
 ### 2.1.7
 * an exception is now properly thrown when a record is not found;